restructure jwt payload, separate module for datasets
